requirements.txt
python-telegram-bot==20.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pytz
